Extended the framework to support DataTemplateSelectors.
Now, DataTemplateSelectors can be created using lambda functions in similar manner as the ValueConverters and MultiValueConverters.
Unlike the converters, the selectors have no interfaces to implement against. Hence, in the implementation the created classes had to derive from the DataTemplateSelector base class. However, this is only an internal implementation detail and won't bother the consumers of this framework too much as it will make minor differences in usage compared to the converters.

Signed-off-by: Dima Enns <dieter.enns@gmail.com>